Implemented new Checker Consume which now outperforms the Regex equivalent

+++ Stringier.Patterns/Alternator.cs
@@ -17,7 +17,7 @@
 		/// <summary>
 		/// The maximum length possibly matched by this pattern
 		/// </summary>
-		internal override Int64 MaxLength => Left.MaxLength >= Right.MaxLength ? Left.MaxLength : Right.MaxLength;
+		internal override Int32 MaxLength => Left.MaxLength >= Right.MaxLength ? Left.MaxLength : Right.MaxLength;
 
 		/// <summary>
 		/// Attempt to consume the <see cref="Alternator"/> from the <paramref name="Source"/>, adjusting the position in the <paramref name="Source"/> as appropriate
