+++ Stringier.Patterns/Optor.cs
@@ -16,7 +16,7 @@ namespace System.Text.Patterns {
 		/// <summary>
 		/// The maximum length possibly matched by this pattern
 		/// </summary>
-		internal override Int64 MaxLength => Pattern.MaxLength;
+		internal override Int32 MaxLength => Pattern.MaxLength;
 
 		/// <summary>
 		/// Attempt to consume the <see cref="Pattern"/> from the <paramref name="Source"/>, adjusting the position in the <paramref name="Source"/> as appropriate
